**Ver1.5.1** May.7th.2020
* Highly optimized the performance of calling customized insertion sequences (non-humman genomes, non-MEIs)!!
* Sample bam file added, example updated, results from sample bam updated!!
* A fatal bug fixed calling L1NE-1 since Ver1.4.1.
* A fatal bug fixed related to the environment of computing clusters and the version of BLASTn. Now require the version ncbi-blast++/2.10.0.
* Minor bugs fixed.
