+++ example/sample_TSD_reads.txt
@@ -1,0 +1,2 @@
+cluster_id	read_name.info	5'_TSD	3'_TSD	Predicted_transD	Unique_26mer_at_5'junction	Whole_insertion_seq
+cluster5_chr19_4879460_4879460_4879460_4879460	SRR3658432.144260_length=32698.5803.6058.2655.2909.4879460.4879460.chr19.+.0.0.0.0.0.0.0	ACAAAA	ACAAAA	ACAAAAACAAATTAGGCCAGGTGTTGGTGGTACATGTCTGTATCCCAGCACTTGGAGAGGCTGAGGCAAAGGAGAATGCTTGAACCCAGGAGGCAGAGTTCTGCACTCGGCTTGGGCACAAAGGAGACTGTCTCAAAAAAAAAAAGGAAGAATGGTTTGTGGCTTGAATAGTAACAATATGACTAGGTATCACATTCTTGGGCCCTCATTTGAGGCATGACACTGTCTTAAGCCAGTTACCTGCTCATCTCATGTAGTCCAGACAGCAAAACCCTGAGGGAAAGGACGATACACTTACCAACTGCCCAGGAAAACTGGAATCAGAGAGGTTCATGATTTGCCAAGTCACACAGCAACTCACTGCAGAGCCAAGGACTCAAGAAACCAGTTCAACCAAACCGGTGATGGTTAACCACTGTGATTACTCACTATTGCCTTGCGCTAAACTCTGACAGTTGGCTGATAATGAACGACCCCCAGGATGGGAAATGTTGTATGTATGTATGTATGAGTGTATGTAGTAGATGTTTTTTGAGACAGATGTTCTCACTCTGTCGACAAGGGTGGAGTGCAGTTGTGTGATCTACGGCTCACTGCAACCTCCATCTCCCAGGTTCAATCGGATTCCTCCTGGCTCAGTCTCCGATTGCTGGTTACAGGCGCGTGCCACAAGGCTGGCTAATTTCTTAATTTTTAGTAGAGACGGGTTTCGCCATATTGTCCAGGGCTGGTCTCATAACTCCTACCAATTCGATCTGCCTGCATCGCTCCCAAAGTGCTGGGATTAAGCGTGAGCCACGCGCATAACTGTACAGGAGGGGTGGGTTCCTTCTCTCGGCAAGGCCATTCCTGGGAGTACTGAAGAGTGTCTGACAGCAGCATCCCCGGCCCCCACCCCTCCATATTGCCAGGAGCGAAACCCCCAGTCATACAACCACAAATGCCCCAGACATCACCCCATGGGCCACTGAAGACAGACTCACCCGCTGGTTTGAGATCTACACTGTTACAGTATTATACTGCGAGGCCTCTGTGGAGTTAATTTCTGATGGTAAATGCTGATGCTGAGAATCCACTTCTGCATGTTTATTATTATTATTATTATTATTATTGTTTTGACAATGAGGTTGGTACGTCCAGTCTCAGGCTGGAGTCAGGGCGGATCATAGCTGCATCAACCTCAAGCTCCTGGGCTCAAACGATCTGCCAACCTCAGCTTCAGAAGTGCTGAGGATTATAACGATGAGCCATGGCGCCCGGCTCTGCTGTTTTGTTGTTTGGTGTTGTTGTTGTTAATTGTTGTTGTCTTTGGGACAGGGTCTTCTGTCTGCCCAAGGCGTGAATGCCAGTGGCGCCCATCTCTGGCTCACCGGCAACTCACCCCTCCCTGGTTCAGAAATTCTCATGCCTCGCCTCCCAGTAGCTGGGAATCTCTCTCTTTTCCTCCTCCTCACGTGTGTTGAGAGAGATATCCCAGGCACTTGGAGGCCGAGGCAATGAGAATGTGCTTAACCCAGGGCAGGGGGGAAGGTTGCGGTGAGCCAAAATGCGCCACTTGCATTCCAGCCTTGGCAAGACAGCAAGACTCTGTCCCAAACAACCAAACATAAAACAACAACACAACAACA	TAAAAATACAAAAATTTATAGGTGGG	ATTTATAGGTATAGGTGGGAATTGAAATAGAGACACTGGACACAGGAAGAGAATCATCACACACCCCGGCCTGATACTGGGGAGGGAGGGGGAGGGAATAGCATGGAGCATACCCTAATGTAAATGACGCAGTTAATGGGTGCAGCACACCAACATGGACTGTTATACATATTGTAACAAAACCTCGCACGTTGTGCACATGTACCTAGAATTTAAGTATAATAATAAAAAAAAAAAAGAACTGAAACATTGAGAAAAACAAAAACAAAAACAAATTAGGCCAGGTGTTGGTGGTACATGTCTGTATCCCAGCACTTGGAGAGGCTGAGGCAAAGGAGAATGCTTGAACCCAGGAGGCAGAGTTCTGCACTCGGCTTGGGCACAAAGGAGACTGTCTCAAAAAAAAAAAGGAAGAATGGTTTGTGGCTTGAATAGTAACAATATGACTAGGTATCACATTCTTGGGCCCTCATTTGAGGCATGACACTGTCTTAAGCCAGTTACCTGCTCATCTCATGTAGTCCAGACAGCAAAACCCTGAGGGAAAGGACGATACACTTACCAACTGCCCAGGAAAACTGGAATCAGAGAGGTTCATGATTTGCCAAGTCACACAGCAACTCACTGCAGAGCCAAGGACTCAAGAAACCAGTTCAACCAAACCGGTGATGGTTAACCACTGTGATTACTCACTATTGCCTTGCGCTAAACTCTGACAGTTGGCTGATAATGAACGACCCCCAGGATGGGAAATGTTGTATGTATGTATGTATGAGTGTATGTAGTAGATGTTTTTTGAGACAGATGTTCTCACTCTGTCGACAAGGGTGGAGTGCAGTTGTGTGATCTACGGCTCACTGCAACCTCCATCTCCCAGGTTCAATCGGATTCCTCCTGGCTCAGTCTCCGATTGCTGGTTACAGGCGCGTGCCACAAGGCTGGCTAATTTCTTAATTTTTAGTAGAGACGGGTTTCGCCATATTGTCCAGGGCTGGTCTCATAACTCCTACCAATTCGATCTGCCTGCATCGCTCCCAAAGTGCTGGGATTAAGCGTGAGCCACGCGCATAACTGTACAGGAGGGGTGGGTTCCTTCTCTCGGCAAGGCCATTCCTGGGAGTACTGAAGAGTGTCTGACAGCAGCATCCCCGGCCCCCACCCCTCCATATTGCCAGGAGCGAAACCCCCAGTCATACAACCACAAATGCCCCAGACATCACCCCATGGGCCACTGAAGACAGACTCACCCGCTGGTTTGAGATCTACACTGTTACAGTATTATACTGCGAGGCCTCTGTGGAGTTAATTTCTGATGGTAAATGCTGATGCTGAGAATCCACTTCTGCATGTTTATTATTATTATTATTATTATTATTGTTTTGACAATGAGGTTGGTACGTCCAGTCTCAGGCTGGAGTCAGGGCGGATCATAGCTGCATCAACCTCAAGCTCCTGGGCTCAAACGATCTGCCAACCTCAGCTTCAGAAGTGCTGAGGATTATAACGATGAGCCATGGCGCCCGGCTCTGCTGTTTTGTTGTTTGGTGTTGTTGTTGTTAATTGTTGTTGTCTTTGGGACAGGGTCTTCTGTCTGCCCAAGGCGTGAATGCCAGTGGCGCCCATCTCTGGCTCACCGGCAACTCACCCCTCCCTGGTTCAGAAATTCTCATGCCTCGCCTCCCAGTAGCTGGGAATCTCTCTCTTTTCCTCCTCCTCACGTGTGTTGAGAGAGATATCCCAGGCACTTGGAGGCCGAGGCAATGAGAATGTGCTTAACCCAGGGCAGGGGGGAAGGTTGCGGTGAGCCAAAATGCGCCACTTGCATTCCAGCCTTGGCAAGACAGCAAGACTCTGTCCCAAACAACCAAACATAAAACAACAACACAACAACA
