+++ scp/7_FP_ex.cpp
@@ -20,8 +20,8 @@ using namespace std;
 
 int fp_ex(string WD_dir, string fasta, string chr, string t, int tsd_index){
     
-    std::ios::sync_with_stdio(false);
-    std::cin.tie(0);
+    //std::ios::sync_with_stdio(false);
+    //std::cin.tie(0);
     
     int BIN_5=50;
     int J_BIN=50;
